pyproject.toml
[project]
name = "back-risk-move"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "joblib>=1.5.3",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "shap>=0.51.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn>=0.46.0",
    "xgboost>=3.2.0",
]
